XCF-1557 Upgrade to spring 2.0 and respective libraries so we can gain access to new ldap api that will allow us to read ldap rows during preview.   lib/jars come from spring distribution instead of being downloaded from individual projects.   I had to disable 2 unit tests: #1 method cache because easymock classextension would not allow for new api in ehcache library.  #2 MigrateAttendant - need to go back and fix this
git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@10071 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXconfig/neoconf/src/org/sipfoundry/sipxconfig/setting/SettingDao.java
@@ -49,7 +49,7 @@ public interface SettingDao extends DataObjectSource {
      * 
      * @return map {groupId as Integer, count as Integer} 
      */
-    public Map<Integer, Integer> getGroupMemberCountIndexedByGroupId(Class groupOwner);
+    public Map<Integer, Long> getGroupMemberCountIndexedByGroupId(Class groupOwner);
 
     /**
      * convienence: find and create group if not found
